refactor(inference): split provider clients into per-provider modules (#1200)
TD-034. Behaviour-preserving decomposition of the two oversized provider
client files into directories, matching the crate CLAUDE.md target shape.

- anthropic_client.rs (1,437 LOC) -> anthropic_client/{mod,wire,json_isolation,sse}.rs
  wire: request/response schema + protocol constants + extract helpers;
  json_isolation: <caller_system>/<engine_instruction> wrapping + structural-tag
  hardening (#458/#599); sse: Anthropic SSE event parser; mod.rs: AnthropicClient
  + generate* methods.
- openai_client.rs (1,160 LOC) -> openai_client/{mod,wire,sse}.rs
  wire: request/response types + public GenerateParams/ResponseFormat/JsonSchemaSpec
  (re-exported so openai_client::* paths are unchanged); sse: streaming read loop
  + line parser; mod.rs: OpenAiClient + build_client_or_fallback.

Public API identical (lib.rs re-exports unchanged). All tests moved with their
subjects. 313 inference tests pass; clippy -D warnings clean; arch_fitness green.
